#211 [refactor] : 패키지를 분리한다
