+++ src/main/java/side/onetime/global/config/SecurityConfig.java
@@ -14,6 +14,7 @@ import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import side.onetime.auth.handler.OAuthLoginFailureHandler;
 import side.onetime.auth.handler.OAuthLoginSuccessHandler;
+import side.onetime.global.filter.JwtFilter;
 
 import java.util.Arrays;
