+++ src/main/java/side/onetime/global/filter/JwtFilter.java
@@ -1,4 +1,4 @@
-package side.onetime.global.config;
+package side.onetime.global.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
